Move pos handler to chunker
