+++ src/main/java/ru/cs/eatright/nlp/RusPosAnnotator.java
@@ -1,4 +1,4 @@
-package ru.cs.eatright.nlpmodel;
+package ru.cs.eatright.nlp;
 
 import java.util.Collections;
 import java.util.List;
