+++ src/main/java/ru/cs/eatright/nlp/StopWordTokenizer.java
@@ -1,4 +1,4 @@
-package ru.cs.eatright.nlpmodel;
+package ru.cs.eatright.nlp;
 
 import org.apache.lucene.analysis.CharArraySet;
 import org.apache.lucene.analysis.TokenStream;
@@ -7,7 +7,7 @@ import org.apache.lucene.analysis.ru.RussianAnalyzer;
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import ru.cs.eatright.nlpmodel.signatures.Token;
+import ru.cs.eatright.nlp.signatures.Token;
 
 import java.io.IOException;
 import java.util.ArrayList;
