+++ src/main/java/ru/cs/eatright/nlp/Tokenizer.java
@@ -1,6 +1,6 @@
-package ru.cs.eatright.nlpmodel;
+package ru.cs.eatright.nlp;
 
-import ru.cs.eatright.nlpmodel.signatures.Token;
+import ru.cs.eatright.nlp.signatures.Token;
 
 import java.io.IOException;
 import java.util.List;
