+++ src/main/resources/applicationContext.xml
@@ -16,13 +16,13 @@
         <constructor-arg name="requestProcessor" ref="requestProcessor"/>
     </bean>
 
-    <bean id="tokenizer" class="ru.cs.eatright.nlpmodel.StopWordTokenizer"/>
+    <bean id="tokenizer" class="ru.cs.eatright.nlp.StopWordTokenizer"/>
     <bean id="requestProcessor" class="ru.cs.eatright.parsing.RequestProcessor">
         <constructor-arg name="tokenizer" ref="tokenizer"/>
     </bean>
-    <bean id="posTagHandler" class="ru.cs.eatright.nlpmodel.PosTagHandler"/>
-    <bean id="chunker" class="ru.cs.eatright.nlpmodel.Chunker"/>
-    <bean id="stemmer" class="ru.cs.eatright.nlpmodel.Stemmer"/>
+    <bean id="posTagHandler" class="ru.cs.eatright.nlp.PosTagHandler"/>
+    <bean id="chunker" class="ru.cs.eatright.nlp.Chunker"/>
+    <bean id="stemmer" class="ru.cs.eatright.nlp.Stemmer"/>
     <bean id="pipeline" class="ru.cs.eatright.parsing.QueryPipeline"/>
 
 </beans>
